第一次大作业/requirements.txt
numpy
opencv-python
jittor
pymaxflow
scipy
pyamg